requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pyyaml==6.0.2
pandas==2.2.2
numpy==1.26.4
sentence-transformers==2.7.0
scipy==1.11.4
faiss-cpu==1.8.0.post1
scikit-learn==1.5.1
# Optional local LLM backends (install manually if needed)
# llama-cpp-python==0.2.84  # Requires CMake + MSVC build tools on Windows; skip unless using llama.cpp
# gpt4all==2.7.0

# Utils
tqdm==4.66.4
loguru==0.7.2
requests==2.32.3
